added server side validation

+++ routes/ideas.js
@@ -56,17 +56,26 @@ router.post("/", async (req, res) => {
 // Update the Idea
 router.put("/:id", async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+
+      return res.json({ success: true, data: updatedIdea });
+    }
+    // Usernames do not Match
+    res.status(403).json({
+      success: false,
+      message: "You are not Authorized to Update this Resource",
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({
@@ -80,8 +89,17 @@ router.put("/:id", async (req, res) => {
 // Deleting the Idea
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Usernames do not Match
+    res.status(403).json({
+      success: false,
+      message: "You are not Authorized to Delete this Resource",
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({
